# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "SWAGGER API DOCUMENTATION",
    "version": "1.0.0",
    "title": "SIMPLE API DOC"
  },
  "host": "localhost:9000",
  "basePath": "/",
  "schemes": ["https", "http"],
  "paths": {
    "/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Users can register",
        "description": "this endpoint uses get request to create users",
        "operationId": "register",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "description": "users email",
            "required": true,
            "type": "string"
          },
          {
            "name": "username",
            "in": "formData",
            "description": "users username",
            "required": true,
            "type": "string"
          },
          {
            "name": "firstname",
            "in": "formData",
            "description": "user firstname",
            "required": true,
            "type": "string"
          },
          {
            "name": "lastname",
            "in": "formData",
            "description": "user lastname",
            "required": true,
            "type": "string"
          },
          {
            "name": "phone",
            "in": "formData",
            "description": "user phone",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "user password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/register/account-activate": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Send Activation Email",
        "description": "This endpoint is used to activate user account",
        "operationId": "AccountActivation",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "send_email",
            "in": "formData",
            "description": "Query params",
            "required": false,
            "type": "number"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "users email",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        }
      }
    },
    "register/account-activate/{hash}": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Activate User Account",
        "description": "Activate user account from link recieved from email",
        "operationId": "ActivateAccount",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "description": "Account Activation Hash",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User Login",
        "description": "this endpoint allows users to login to their account",
        "operationId": "Login",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "description": "user email",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "user password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "get": {
        "tags": ["Accounts"],
        "summary": "Returns all accounts",
        "description": "Staffs and admin can see all accounts",
        "operationId": "getaccounts",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  }
}